requirements.txt
# Core Data Processing
pandas==2.1.4
numpy==1.26.2

# Database
sqlite3  # Built-in with Python

# Data Visualization
plotly==5.18.0
streamlit==1.29.0

# Google Gemini AI
google-generativeai==0.3.2
google-ai-generativelanguage==0.4.0

# Image Processing
Pillow==10.1.0

# Web Scraping (if needed)
beautifulsoup4==4.12.2
requests==2.31.0
lxml==4.9.3

# Data Analysis
scipy==1.11.4
scikit-learn==1.3.2

# Environment Variables
python-dotenv==1.0.0

# Jupyter (for EDA notebooks)
jupyter==1.0.0
notebook==7.0.6

# Utilities
openpyxl==3.1.2  # Excel file support
python-dateutil==2.8.2

# Testing (optional)
pytest==7.4.3

# Code Quality (optional)
black==23.12.1
flake8==6.1.0
